✨ Restore performance timeline commands from master
- Uncommented and properly implemented get_performance_timeline functions
- get_performance_timeline uses performance_tracker.get_current_timeline()
- get_performance_timeline_for_transcript uses database method
- Added overlay module to commands
- These functions were working in master and just needed proper migration
